Moving struct definitions to mod.rs files
