+++ src/configuration/config.rs
@@ -1,17 +1,7 @@
-use super::parser::{Cli, Mode};
-use clap::Parser;
-
+use super::Config;
+use crate::configuration::Cli;
 use crate::game::elements::Point;
-
-#[derive(Debug)]
-pub struct Config {
-    pub size: Point,
-    pub char: char,
-    pub fps: u64,
-    pub mode: Mode,
-    pub rand_points: usize,
-    pub duration: Option<usize>,
-}
+use clap::Parser;
 
 impl Config {
     pub fn configure() -> Config {
